% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_breaks.R
\name{add_breaks}
\alias{add_breaks}
\title{A function to add breaks on a plot}
\usage{
add_breaks(
  density_plot = density_plot,
  legendbreaks = legendbreaks,
  legendlimits = legendlimits,
  legendlabels = legendlabels
)
}
\arguments{
\item{density_plot}{is the a ggplot}

\item{legendbreaks}{is the breaks in the legend}

\item{legendlimits}{are the limits in the legend}

\item{legendlabels}{are the labels in the legend important when you want to use symbols}
}
\value{
a ggplot with a legend label in the color of light blue to red
}
\description{
A function to add breaks on a plot
}
\examples{
add_breaks(density_plot=density_plot,
legendbreaks=c(0,0.5,1,1.5,2.0),legendlimits=c(0,2),
legendlabels=c('     0','   0.5','   1.0','   1.5','> 2.0'))
}
