% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend.R
\name{add_legend}
\alias{add_legend}
\title{A function to add the legend inside the plot}
\usage{
add_legend(
  plot_wbreaks = plot_wbreaks,
  legxy = legxy,
  legtx = legtx,
  legty = legty,
  xval = xval,
  yval = yval
)
}
\arguments{
\item{plot_wbreaks}{is the plot already with selected breaks in the colors}

\item{legxy}{is the place where the legend will be set inside the panel}

\item{legtx}{is where the title for the legend will be set in the x axis}

\item{legty}{is where the title for the legend will be set in the y axis}

\item{xval}{are the limits from the x axis}

\item{yval}{are the limits for the y axis}
}
\value{
a plot with the legend inside a defined limits
}
\description{
A function to add the legend inside the plot
}
\examples{
add_legend(plot_wbreaks=plot_wbreaks,legtx=3905000,
legty=3510000,legxy=c(0.11, 0.21),xval=c(3910000,4250000),yval=c(3380000,3680000))
}
