% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_theme.R
\name{add_theme}
\alias{add_theme}
\title{A function to add my theme to a ggplot}
\usage{
add_theme(plot_wlegend = plot_wlegend)
}
\arguments{
\item{plot_wlegend}{is the plot to be modified}
}
\value{
a plot with specific colors and lines around, a little bit like theme_bw but with latitude and longitude
}
\description{
A function to add my theme to a ggplot
}
\examples{
add_theme(plot_wlegend = density_wmap)
}
