% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basis_test.R
\docType{data}
\name{Basis_test}
\alias{Basis_test}
\title{Basis_test contains Basis Information}
\format{
Data frame with 74440 obs and 22 columns
}
\usage{
data(Basis_test)
}
\description{
Basis_test contains Basis Information
}
\examples{
data(Basis_test)
}
\references{
DDA
}
\keyword{datasets}
