% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasisDaten_test.R
\docType{data}
\name{BasisDaten_test}
\alias{BasisDaten_test}
\title{BasisDaten_test contains BasisDaten Information}
\format{
Data frame with 74440 obs and 22 columns
}
\usage{
data(BasisDaten_test)
}
\description{
BasisDaten_test contains BasisDaten Information
}
\examples{
data(BasisDaten_test)
}
\references{
DDA
}
\keyword{datasets}
