% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_taxonomicgroup.R
\name{classify_taxonomic}
\alias{classify_taxonomic}
\title{Load taxonomic groups}
\usage{
classify_taxonomic(my_data = my_data, col_name = col_name)
}
\arguments{
\item{my_data}{a data frame that includes a column with english estandar names}

\item{col_name}{the name of the column}
}
\value{
taxonomic groups to classify species
}
\description{
Load taxonomic groups
}
\examples{
classify_taxonomic(my_data=data.frame(Species=c('Black-throated Diver')),col_name = "Species")
}
