% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code_Descriptions.R
\docType{data}
\name{Code_Descriptions}
\alias{Code_Descriptions}
\title{Descriptions include information for filling up the columns}
\format{
Data frame with 70 obs and 18 variables
}
\usage{
data(Code_Descriptions)
}
\description{
Descriptions include information for filling up the columns
}
\examples{
data(Code_Descriptions)
}
\references{
DDA
}
\keyword{datasets}
