% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Code_Euring.R
\docType{data}
\name{Code_Euring}
\alias{Code_Euring}
\title{Euroring codes and artificial taxonomic classificacions}
\format{
Data frame with 549 obs and columns code, scientific name, english name and artififical taxonomical classification
}
\usage{
data(Code_Euring)
}
\description{
Euroring codes and artificial taxonomic classificacions
}
\examples{
data(Code_Euring)
}
\references{
https://euring.org/
}
\keyword{datasets}
