% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Column_Descriptions.R
\docType{data}
\name{Column_Descriptions}
\alias{Column_Descriptions}
\title{Includes column names and explanation for filling them}
\format{
Data frame with column names and explanations
}
\usage{
data(Column_Descriptions)
}
\description{
Includes column names and explanation for filling them
}
\examples{
data(Column_Descriptions)
}
\references{
DDA
}
\keyword{datasets}
