% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities_3035.R
\docType{data}
\name{densities_3035}
\alias{densities_3035}
\title{Random generated data
2000 observations from eleven variables.}
\format{
Data frame with random generated observations and eleven variables
}
\usage{
data(densities_3035)
}
\description{
Random generated data
2000 observations from eleven variables.
}
\examples{
data(densities_3035)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
