% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_df.R
\docType{data}
\name{density_df}
\alias{density_df}
\title{Random generated data
50 observations from 10 variables.}
\format{
Data frame with random generated observations
}
\usage{
data(density_df)
}
\description{
Random generated data
50 observations from 10 variables.
}
\examples{
data(density_df)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
