% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_grid.R
\docType{data}
\name{density_grid}
\alias{density_grid}
\title{Random generated data
251 observations from three variables.}
\format{
Simple features (sf) with random generated observations in a grid
}
\usage{
data(density_grid)
}
\description{
Random generated data
251 observations from three variables.
}
\examples{
data(density_grid)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
