% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\docType{data}
\name{density_plot}
\alias{density_plot}
\title{ggplot list of 9}
\format{
A ggplot with the densities and backgroun map
}
\usage{
data(density_plot)
}
\description{
ggplot list of 9
}
\examples{
data(density_plot)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
