% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_survey.R
\docType{data}
\name{density_survey}
\alias{density_survey}
\title{Random generated data
2000 observations from ten variables.}
\format{
Data frame with random generated observations and nine variables
}
\usage{
data(density_survey)
}
\description{
Random generated data
2000 observations from ten variables.
}
\examples{
data(density_survey)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
