% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_surveyed.R
\docType{data}
\name{grid_surveyed}
\alias{grid_surveyed}
\title{Contains a sf object with a grid from the North Sea of 5 x 5 in is 3035 CRS
A total of 251 observations of 4 variables}
\format{
A grid 5x5 of the north sea
}
\usage{
data(grid_surveyed)
}
\description{
Contains a sf object with a grid from the North Sea of 5 x 5 in is 3035 CRS
A total of 251 observations of 4 variables
}
\examples{
data(grid_surveyed)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
