% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_basemap.R
\name{load_basemap}
\alias{load_basemap}
\title{A function to create basemap}
\usage{
load_basemap(
  my_CRS = my_CRS,
  color_land = color_land,
  color_water = color_water
)
}
\arguments{
\item{my_CRS}{define your CRS, for example 4326}

\item{color_land}{select color in hex format for land}

\item{color_water}{select color in hex format for water}
}
\value{
returns a ggplot
}
\description{
A function to create basemap
}
\examples{
load_basemap(my_CRS=4326, color_land='#f7bf54', color_water='#3668b4')
}
