% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_survey.R
\docType{data}
\name{my_survey}
\alias{my_survey}
\title{Random generated data
2000 observations from nine variables.}
\format{
Data frame with random generated observations and nine variables
}
\usage{
data(my_survey)
}
\description{
Random generated data
2000 observations from nine variables.
}
\examples{
data(my_survey)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
