% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{A function to plot densities inside Europe, focused on the North Sea}
\usage{
plot_density(
  density_grid = density_grid,
  column_density = column_density,
  color_land = color_land,
  color_water = color_water
)
}
\arguments{
\item{density_grid}{a grid with density information}

\item{column_density}{the column where the density is stored}

\item{color_land}{the color of the land shapefile}

\item{color_water}{the color of the sea shapefile}
}
\value{
a plot with densities plotted
}
\description{
A function to plot densities inside Europe, focused on the North Sea
}
\examples{
plot_density(density_grid=density_grid,column_density='mean_density',
color_land='#f7bf54', color_water='#3668b4')
}
