% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wbreaks.R
\docType{data}
\name{plot_wbreaks}
\alias{plot_wbreaks}
\title{A plot
Create density plot in ggplot2.}
\format{
List of 9 (ggplot format)
}
\usage{
data(plot_wbreaks)
}
\description{
A plot
Create density plot in ggplot2.
}
\examples{
data(plot_wbreaks)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
