% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_density.R
\name{subset_density}
\alias{subset_density}
\title{A function to add density values to the grid, and keep only the grid will values.}
\usage{
subset_density(
  density_survey = density_survey,
  column_density = column_density,
  grid_identifier = grid_identifier,
  survey_grid = survey_grid
)
}
\arguments{
\item{density_survey}{a sf object with geometry and density columns}

\item{column_density}{the name of the column where the density values are stored}

\item{grid_identifier}{is the name that each grid has}

\item{survey_grid}{is the grid from to the surveyed area}
}
\value{
a sf object with the surveyed grid and in each grid the average densities
}
\description{
A function to add density values to the grid, and keep only the grid will values.
}
\examples{
subset_density(density_survey=densities_3035,column_density='densities',
survey_grid=grid_surveyed,grid_identifier='grid_id')
}
