% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_grid.R
\name{subset_grid}
\alias{subset_grid}
\title{A function to subset the grid and keep only grid cells that were sampled}
\usage{
subset_grid(
  survey_grid = survey_grid,
  grid_identifier = grid_identifier,
  survey_data = survey_data
)
}
\arguments{
\item{survey_grid}{the grid as sf object}

\item{grid_identifier}{the name inside each grid to be able to tell them apart}

\item{survey_data}{the data from your survey as sf object}
}
\value{
a clean grid, only grids with values are returned
}
\description{
A function to subset the grid and keep only grid cells that were sampled
}
\examples{
grid_clean<-subset_grid(survey_grid=grid5x5_3035,survey_data=survey_3035,grid_identifier='grid_id')
}
