% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_3035.R
\docType{data}
\name{survey_3035}
\alias{survey_3035}
\title{Random generated data
2000 observations from nine variables.}
\format{
Data frame with random generated observations and nine variables
}
\usage{
data(survey_3035)
}
\description{
Random generated data
2000 observations from nine variables.
}
\examples{
data(survey_3035)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
