% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_4326.R
\docType{data}
\name{survey_4326}
\alias{survey_4326}
\title{Random generated data
2000 observations from nine variables.}
\format{
Data frame with 3 variables
}
\usage{
data(survey_4326)
}
\description{
Random generated data
2000 observations from nine variables.
}
\examples{
data(survey_4326)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
