% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_densities.R
\docType{data}
\name{survey_densities}
\alias{survey_densities}
\title{Random generated data
2000 observations from ten variables.}
\format{
Data frame with random generated observations and nine variables
}
\usage{
data(survey_densities)
}
\description{
Random generated data
2000 observations from ten variables.
}
\examples{
data(survey_densities)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
