% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_survey.R
\name{transform_survey}
\alias{transform_survey}
\title{A function to transform the CRS from the dataframe and coverts it to sf object}
\usage{
transform_survey(
  survey_data = survey_data,
  column_longitude = column_longitude,
  column_latitude = column_latitude,
  from_CRS = from_CRS,
  to_CRS = to_CRS
)
}
\arguments{
\item{survey_data}{your data to transform}

\item{column_longitude}{the name of the column where the longitude data is}

\item{column_latitude}{the name of the column where the latitude data is}

\item{from_CRS}{the CRS of the data collection for example degrees is in 4236}

\item{to_CRS}{the CRS of the final data neede for example 3035}
}
\value{
a sf object with an assigned CRS
}
\description{
A function to transform the CRS from the dataframe and coverts it to sf object
}
\examples{
survey_3035<-transform_survey(survey_data=survey_4326,
column_latitude='latitude',column_longitude='longitude',
from_CRS=4326,to_CRS=3035)
}
