% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trip_test.R
\docType{data}
\name{Trip_test}
\alias{Trip_test}
\title{Trip_test contains Trip Information}
\format{
Data frame with 1 obs and 36 columns
}
\usage{
data(Trip_test)
}
\description{
Trip_test contains Trip Information
}
\examples{
data(Trip_test)
}
\references{
DDA
}
\keyword{datasets}
