% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TripDaten_test.R
\docType{data}
\name{TripDaten_test}
\alias{TripDaten_test}
\title{TripDaten_test contains TripDaten Information}
\format{
Data frame with 1 obs and 36 columns
}
\usage{
data(TripDaten_test)
}
\description{
TripDaten_test contains TripDaten Information
}
\examples{
data(TripDaten_test)
}
\references{
DDA
}
\keyword{datasets}
