% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend.R
\name{add_legend}
\alias{add_legend}
\title{A function to add the legend inside the plot}
\usage{
add_legend(plot_wbreaks = plot_wbreaks, legxy = legxy)
}
\arguments{
\item{plot_wbreaks}{is the plot already with selected breaks in the colors}

\item{legxy}{is the place where the legend will be set inside the panel}
}
\value{
a plot with the legend inside a defined limits
}
\description{
A function to add the legend inside the plot
}
\examples{
add_legend(plot_wbreaks=density_wmap,legxy=c(0.11, 0.21))
}
