% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_survey.R
\name{transform_survey}
\alias{transform_survey}
\title{A function to transfor the CRS from your data frame}
\usage{
transform_survey(
  survey_data = survey_data,
  column_longitude = column_longitude,
  column_latitude = column_latitude,
  from_CRS = from_CRS,
  to_CRS = to_CRS
)
}
\arguments{
\item{survey_data}{A data frame with longitude and latitude in degrees}

\item{column_longitude}{The name of the column where the longitude data is}

\item{column_latitude}{The name of the column where the latitude data is}

\item{from_CRS}{The CRS when the data was collected, if is degrees probably 4326}

\item{to_CRS}{The CRS that you want your data to be presented, for example 3035}
}
\value{
returns a sf object with CRS
}
\description{
A function to transfor the CRS from your data frame
}
\examples{
survey_3035<-transform_survey(survey_data=survey_4326,
column_latitude='latitude',column_longitude='longitude',
from_CRS=4326,to_CRS=3035)
}
